[INDY-2154] flake8
